+++ front/app/containers/ProjectsShowPage/shared/header/ProjectInfoSideBar.tsx
@@ -15,6 +15,9 @@ import usePhases from 'hooks/usePhases';
 import useEvents from 'hooks/useEvents';
 import useAuthUser from 'hooks/useAuthUser';
 
+// router
+import clHistory from 'utils/cl-router/history';
+
 // services
 import { IPhaseData, getCurrentPhase, getLastPhase } from 'services/phases';
 
@@ -172,6 +175,9 @@ const ProjectInfoSideBar = memo<Props>(({ projectId, className }) => {
   }, []);
 
   if (!isNilOrError(project)) {
+    const postingIsEnabled =
+      project.attributes.posting_enabled ||
+      currentPhase?.attributes.posting_enabled;
     const projectType = project.attributes.process_type;
     const projectParticipantsCount = project.attributes.participants_count;
     const maxBudget =
@@ -365,6 +371,38 @@ const ProjectInfoSideBar = memo<Props>(({ projectId, className }) => {
                 )}
               </ListItem>
             )}
+            {((projectType === 'continuous' &&
+              projectParticipationMethod === 'native_survey') ||
+              currentPhaseParticipationMethod === 'native_survey') &&
+              postingIsEnabled && (
+                <ListItem>
+                  <ListItemIcon ariaHidden name="survey" />
+                  {!isNilOrError(authUser) ? (
+                    <ListItemButton
+                      id="e2e-project-sidebar-surveys-count"
+                      onClick={() => {
+                        clHistory.push(
+                          `/projects/${project.attributes.slug}/ideas/new`
+                        );
+                      }}
+                    >
+                      <FormattedMessage
+                        {...(projectType === 'continuous'
+                          ? messages.xSurveys
+                          : messages.xSurveysInCurrentPhase)}
+                        values={{ surveysCount: 1 }}
+                      />
+                    </ListItemButton>
+                  ) : (
+                    <FormattedMessage
+                      {...(projectType === 'continuous'
+                        ? messages.xSurveys
+                        : messages.xSurveysInCurrentPhase)}
+                      values={{ surveysCount: 1 }}
+                    />
+                  )}
+                </ListItem>
+              )}
             {((projectType === 'continuous' &&
               projectParticipationMethod === 'poll') ||
               currentPhaseParticipationMethod === 'poll') && (