+++ front/app/modules/commercial/google_tag_manager/index.tsx
@@ -15,6 +15,10 @@ declare module 'components/ConsentManager/destinations' {
   export interface IDestinationMap {
     google_tag_manager: 'google_tag_manager';
   }
+
+  interface IConsentManagerFeatureMap {
+    google_tag_manager: 'google_tag_manager';
+  }
 }
 
 const destinationConfig: IDestinationConfig = {