+++ CHANGELOG.md
@@ -1,6 +1,102 @@
 # Changelog
 
-## Next Release
+### Fixed
+
+- [CL-1586] Fix deleting folder and cause images
+
+## 2022-09-08
+
+### Fixed
+
+- Save button text in representation data input interface is now translated.
+
+## 2022-09-07_2
+
+### Changed
+
+- Custom forms can be associated with projects and phases.
+- Survey form builder improvements.
+
+### 2022-09-07
+
+### Fixed
+
+- Input status dashboard card now updates immediately (not just after refresh) when changing idea status in input manager
+
+## 2022-09-06
+
+### Added
+
+- [CL-423] Added input feedback dashboard card with new analytics endpoint
+
+## 2022-09-05
+
+### Added
+
+- [CL-1157] Added domicile fields to representation dashboard
+
+## 2022-09-02
+
+### Fixed
+- [CL-1580] Do not show full admin panel to moderators
+
+## 2022-08-30
+
+### Fixed
+- [CL-1308] Do not allow editing page slug when custom navbar is disabled
+
+## 2022-08-29
+
+### Fixed
+- [SLS-65] Fixed issue in citizenlab-ee to allow bulk import of custom field options
+
+## 2022-08-24
+
+### Fixed
+- [CL-1505] Posting ideas when there is no current phase
+- [CL-1509] Bulk idea import works for special date cells
+
+## 2022-08-23
+
+### Added
+- The public API now supports phases, as well as additional project properties
+- The project search now also searches through content builder content
+
+## 2022-08-18
+
+### Fixed
+ - [CL-1407] Quickfix for critical bug that made the 'Submit your idea' button disappear for projects with a timeline
+
+### Added
+
+- [CL-1097] Added the ability for users to search for projects and folders by keyword on the main projects index page
+
+### Changed
+
+- Improve error handling and user feedback in the forms for creating pages and custom navigation items
+- [CL-854] Added tooltip to admin budget field in idea form for clarity
+
+## 2022-08-11
+
+### Fixed
+
+- [CL-1302] Added topics column to bulk idea import example sheet
+
+## 2022-08-09
+
+### Fixed
+
+- [CL-1289] Fix 'Submit base data' button width (representation dashboard)
+
+### Added
+
+- [CL-1273] Enable age graphs in representation dashboard
+
+## 2022-08-04
+
+### Added
+
+-[CL-5] Customizable Pages iteration 1
 
 ## 2022-08-03
 
@@ -67,6 +163,7 @@
 
 ### Added
 
+- [CL-1058] - Add desktop preview in content builder
 - [CL-1096] Add representativeness score to header of representativeness chart card
 
 ## 2022-07-07
@@ -99,6 +196,7 @@
 ### Fixed
 
 - [CL-1083] XLSX export issues with custom fields
+
 
 ## 2022-07-04